clear warnings

+++ db/filename.cc
@@ -47,7 +47,7 @@ static std::string MakeFileName2(const Options & options, uint64_t number,
                static_cast<unsigned long long>(number),
                suffix);
 
-  return((level<options.tiered_slow_level ?
+  return((level<(int)options.tiered_slow_level ?
           options.tiered_fast_prefix : options.tiered_slow_prefix) + buf);
 }
 
@@ -61,7 +61,7 @@ std::string MakeDirName2(const Options & options,
       snprintf(buf, sizeof(buf), "/%s",
                suffix);
 
-  return((level<options.tiered_slow_level ?
+  return((level<(int)options.tiered_slow_level ?
           options.tiered_fast_prefix : options.tiered_slow_prefix) + buf);
 }
 
@@ -247,7 +247,7 @@ MakeTieredDbname(
     const std::string & dbname,    // input ... original dbname from DBImpl constructor
     Options & options)             // input/output ... writable Options, tiered values changed
 {
-    if (0<options.tiered_slow_level && options.tiered_slow_level<config::kNumLevels
+    if (0<(int)options.tiered_slow_level && (int)options.tiered_slow_level<config::kNumLevels
         && 0!=options.tiered_fast_prefix.size() && 0!=options.tiered_slow_prefix.size())
     {
         options.tiered_fast_prefix.append("/");
